styling navbar
