+++ src/front/js/component/footer.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { Link } from "react-router-dom";
 
 export const Footer = () => (
 	<div className="container-fluid" style={{backgroundColor: "#040926", color: "#E0E1DD"}}>
@@ -8,7 +9,7 @@ export const Footer = () => (
       			<li className="nav-item"><a href="#" className="nav-link px-2 text-muted">Perdidas</a></li>
       			<li className="nav-item"><a href="#" className="nav-link px-2 text-muted">Encontradas</a></li>
       			<li className="nav-item"><a href="#" className="nav-link px-2 text-muted">Mapa</a></li>
-      			<li className="nav-item"><a href="#" className="nav-link px-2 text-muted">FAQ</a></li>
+      			<li className="nav-item"><Link to="/preguntas-frecuentes" className="nav-link px-2 text-muted">FAQ</Link></li>
 				<li className="nav-item"><a href="#" className="nav-link px-2 text-muted">Contacto</a></li>
     		</ul>
 			<p className="text-center text-muted">Ururguay</p>
