+++ src/front/js/component/navbar.js
@@ -3,6 +3,7 @@ import { Context } from "../store/appContext";
 import { Link, useNavigate, useLocation } from "react-router-dom";
 import logoClaro from "../../img/logoClaro.png"
 import logoClaroGif from "../../img/logo-mascotapp_claro.gif"
+import "../../styles/navbar.css"
 
 export const Navbar = () => {
   const nav = useNavigate();
@@ -12,11 +13,11 @@ export const Navbar = () => {
 
 
   return (
-      <header className="navbar navbar-expand-md navbar-dark sticky-top" style={{ backgroundColor: "#040926", color: "#E0E1DD" }}>
+      <nav className="navbar navbar-expand-md navbar-dark sticky-top" style={{ backgroundColor: "#040926", color: "#E0E1DD" }}>
           <div className="container">
-              <button id="botonPrincipal" className="image-container" onClick={() => nav("/")}>
-                  <img src={logoClaro} alt="Static Image" className="static-image" />
-                  <img src={logoClaroGif} alt="Animated Image" className="animated-image" />
+              <button id="botonPrincipal" className="image-container pb-0" onClick={() => nav("/")}>
+                  <img src={logoClaro} alt="Static Image" className="static-image pb-0" />
+                  <img src={logoClaroGif} alt="Animated Image" className="animated-image pb-0" />
               </button>
               <button className="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNav" aria-controls="navbarNav" aria-expanded="false" aria-label="Toggle navigation">
                   <span className="navbar-toggler-icon"></span>
@@ -61,7 +62,7 @@ export const Navbar = () => {
                   </ul>
               </div>
           </div>
-      </header>
+      </nav>
   );
 };
