+++ src/front/styles/navbar.css
@@ -1,0 +1,4 @@
+.image-container img {
+    width: 70px; /* Ajusta el tamaño según tus necesidades */
+    height: auto; /* Mantiene la proporción */
+  }
